+++ Program.cs
@@ -3,7 +3,7 @@ using Avalonia;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.ReactiveUI;
 
-namespace app
+namespace student_concern_ticketing_system
 {
     class Program
     {
