--- frontend/App.axaml.cs
@@ -1,32 +1,0 @@
-using Avalonia;
-using Avalonia.Controls.ApplicationLifetimes;
-using Avalonia.Markup.Xaml;
-using app.ViewModels;
-using app.Views;
-using app.Services;
-
-namespace app
-{
-    public class App : Application
-    {
-        public override void Initialize()
-        {
-            AvaloniaXamlLoader.Load(this);
-        }
-
-        public override void OnFrameworkInitializationCompleted()
-        {
-            if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
-            {
-                var db  = new Database();
-
-                desktop.MainWindow = new MainWindow
-                {
-                    DataContext = new MainWindowViewModel(db),
-                };
-            }
-
-            base.OnFrameworkInitializationCompleted();
-        }
-    }
-}
