+++ frontend/Models/Ticket.cs
@@ -1,10 +1,10 @@
-namespace app.Models
+namespace student_concern_ticketing_system.frontend.Models
 {
     public class Ticket
     {
         public int id { get; set; }
-        public string TicketName { get; set; }
-        public string Content { get; set; }
+        public string? TicketName { get; set; }
+        public string? Content { get; set; }
         public bool IsFinished { get; set; }
     }
 }
