+++ frontend/Views/MainWindow.axaml
@@ -1,15 +1,15 @@
 <Window xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:vm="using:app.ViewModels"
-        xmlns:views="using:app.Views"
+        xmlns:vm="using:student_concern_ticketing_system.frontend.ViewModels"
+        xmlns:views="using:student_concern_ticketing_system.frontend.Views"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="app.Views.MainWindow"
+        x:Class="student_concern_ticketing_system.frontend.Views.MainWindow"
         Icon="/Assets/avalonia-logo.ico"
         TransparencyLevelHint="AcrylicBlur"
         Background="Transparent"
-        Title="app">
+        Title="Student Concern Ticketing System">
 
    <Design.DataContext>
        <vm:MainWindowViewModel />
