+++ frontend/Views/NavigationSidebar.axaml
@@ -3,7 +3,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="app.Views.NavigationSidebar">
+             x:Class="student_concern_ticketing_system.frontend.Views.NavigationSidebar">
   <StackPanel>
     <Button DockPanel.Dock="Left">Option #1</Button>
     <Button DockPanel.Dock="Left">Option #2</Button>
