+++ frontend/Views/TicketListView.axaml
@@ -3,7 +3,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d" d:DesignWidth="200" d:DesignHeight="400"
-             x:Class="app.Views.TicketListView">
+             x:Class="student_concern_ticketing_system.frontend.Views.TicketListView">
   <DockPanel>
     <Button DockPanel.Dock="Bottom">Add an item</Button>
     <ItemsControl Items="{Binding Items}">
